+++ calcipy/doit_tasks/lint.py
@@ -101,8 +101,7 @@ def _lint_non_python() -> List[DoitAction]:
     """
     actions = []
     pbs = get_dg().meta.paths_by_suffix
-    paths_yaml = pbs.get('yml', []) + pbs.get('yaml', [])
-    if paths_yaml:
+    if paths_yaml := pbs.get('yml', []) + pbs.get('yaml', []):
         paths = ' '.join(f'"{pth}"' for pth in paths_yaml)
         actions.append(Interactive(f'poetry run yamllint {paths}'))
     return actions
@@ -116,7 +115,8 @@ def task_lint_python() -> DoitTask:
         DoitTask: doit task
 
     """
-    actions = _lint_python(get_dg().lint.paths_py, path_flake8=get_dg().lint.path_flake8)
+    dg = get_dg()
+    actions = _lint_python(dg.lint.paths_py, path_flake8=dg.lint.path_flake8)
     return debug_task(actions)
 
 
@@ -186,7 +186,8 @@ def task_static_checks() -> DoitTask:
     """
     paths = ' '.join(map(Path.as_posix, get_dg().lint.paths_py))
     return debug_task([
-        Interactive(f'poetry run vulture {paths} --min-confidence 70 --sort-by-size'),
+        # Set to 61% because pydantic.Config and attributes are flagged with 60% confidence
+        Interactive(f'poetry run vulture {paths} --min-confidence 61 --sort-by-size'),
     ])
 
 
@@ -198,9 +199,36 @@ def task_security_checks() -> DoitTask:
         DoitTask: doit task
 
     """
+    # TODO: Implement semgrep - what are a good ruleset to start with?
+    #   https://github.com/returntocorp/semgrep-rules/tree/develop/python
+    #   https://awesomeopensource.com/project/returntocorp/semgrep-rules?categorypage=45
+    configs = ' '.join([
+        # See more at: https://semgrep.dev/explore
+        '--config=p/ci',
+        '--config=p/security-audit',
+        '--config=r/python.airflow',
+        '--config=r/python.attr',
+        '--config=r/python.click',
+        '--config=r/python.cryptography',
+        '--config=r/python.distributed',
+        '--config=r/python.docker',
+        '--config=r/python.flask',
+        '--config=r/python.jinja2',
+        '--config=r/python.jwt',
+        '--config=r/python.lang',
+        '--config=r/python.pycryptodome',
+        '--config=r/python.requests',
+        '--config=r/python.security',
+        '--config=r/python.sh',
+        '--config=r/python.sqlalchemy',
+        # dlukeomalley:unchecked-subprocess-call
+        # dlukeomalley:use-assertEqual-for-equality
+        # dlukeomalley:flask-set-cookie
+        # clintgibler:no-exec
+    ])
     return debug_task([
         Interactive(f'poetry run bandit --recursive {get_dg().meta.pkg_name}'),
-        Interactive('poetry run nox --session check_security'),
+        Interactive(f'semgrep {get_dg().meta.pkg_name} {configs}'),
     ])