+++ calcipy/file_helpers.py
@@ -4,6 +4,7 @@ import os
 import shutil
 import string
 import time
+from contextlib import suppress
 from pathlib import Path
 
 import yaml
@@ -84,7 +85,7 @@ def get_doc_dir(path_project: Path) -> Path:
 
     """
     path_copier = path_project / _COPIER_ANSWERS_NAME
-    return path_project / _read_yaml_file(path_copier).get('doc_dir', 'docs')
+    return path_project / _read_yaml_file(path_copier).get('doc_dir', 'docs')  # type: ignore[no-any-return]
 
 
 @beartype
@@ -216,7 +217,6 @@ def get_relative(full_path: Path, other_path: Path) -> Optional[Path]:
         relative path
 
     """
-    try:
+    with suppress(ValueError):
         return full_path.relative_to(other_path)
-    except ValueError:
-        return
+    return None