+++ calcipy/proc_helpers.py
@@ -8,7 +8,7 @@ from beartype.typing import Callable, Optional
 
 
 @beartype
-def run_cmd(cmd: str, printer: Optional[Callable[[str], None]] = None, **kwargs) -> str:
+def run_cmd(cmd: str, printer: Optional[Callable[[str], None]] = None, **kwargs) -> str:  # type: ignore[no-untyped-def]
     """Run command with subprocess and return the output.
 
     Inspired by: https://stackoverflow.com/a/38745040/3219667
@@ -16,7 +16,7 @@ def run_cmd(cmd: str, printer: Optional[Callable[[str], None]] = None, **kwargs)
     Args:
         cmd: string command
         printer: optional callable to output the lines in real time
-        kwargs: any additional keyword arguments to pass to `subprocess.Popen` (typically `cwd`)
+        **kwargs: any additional keyword arguments to pass to `subprocess.Popen` (typically `cwd`)
 
     Returns:
         str: stripped output
@@ -29,19 +29,18 @@ def run_cmd(cmd: str, printer: Optional[Callable[[str], None]] = None, **kwargs)
         cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, universal_newlines=True,
         shell=True, **kwargs,  # noqa: S602
     ) as proc:
-        stdout: BufferedReader = proc.stdout  # type: ignore
+        stdout: BufferedReader = proc.stdout  # type: ignore[assignment]
         lines = []
         return_code = None
         while return_code is None:
-            line = stdout.readline()
-            if line:
+            if line := stdout.readline():
                 lines.append(line)
                 if printer:
-                    printer(line.rstrip())
+                    printer(line.rstrip())  # type: ignore[arg-type]
             else:
                 return_code = proc.poll()
 
-    output = ''.join(lines)
+    output = ''.join(lines)  # type: ignore[unreachable]
     if return_code != 0:
         raise subprocess.CalledProcessError(returncode=return_code, cmd=cmd, output=output)
     return output