+++ tests/test_doit_tasks/test_code_tags.py
@@ -1,13 +1,10 @@
 """Test code_tags.py."""
 
-from calcipy.code_tag_collector import write_code_tag_file
 from calcipy.doit_tasks.code_tags import task_collect_code_tags
 
 
-def test_task_collect_code_tags():
+def test_task_collect_code_tags(assert_against_cache):
     """Test task_collect_code_tags."""
     result = task_collect_code_tags()
 
-    actions = result['actions']
-    assert len(actions) == 1
-    assert isinstance(actions[0][0], type(write_code_tag_file))
+    assert_against_cache(result['actions'])