+++ tests/test_dot_dict.py
@@ -2,6 +2,8 @@
 
 import arrow
 import pytest
+from hypothesis import given
+from hypothesis import strategies as st
 
 from calcipy.dot_dict import ddict
 
@@ -17,6 +19,24 @@ from calcipy.dot_dict import ddict
     ],
 )
 def test_ddict(key, value):
+    """Test ddict."""
+    result = ddict(**{key: value})
+
+    assert getattr(result, key) == value
+    assert result[key] == value
+    assert isinstance(result, dict)
+    assert result.get(f'--{key}--') is None
+
+
+_ST_ANY = st.booleans() | st.binary() | st.integers() | st.text()
+"""Broadest swatch of strategies for data input testing of dotdict."""
+
+
+@given(
+    key=st.text(),
+    value=(_ST_ANY | st.dictionaries(keys=_ST_ANY, values=_ST_ANY, max_size=10)),
+)
+def test_ddict_with_hypothesis(key, value):
     """Test ddict."""
     result = ddict(**{key: value})